requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

chromadb==0.5.0
langchain-community==0.3.13
langchain-core==0.3.28
langchain-groq==0.2.2
pandas==2.0.2
python-dotenv==1.0.0
streamlit==1.35.0
